% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chk_num.R
\name{chk_num}
\alias{chk_num}
\title{Checa si las variables son numéricas}
\usage{
chk_num(b, v)
}
\arguments{
\item{b}{Base de datos}

\item{v}{Conjunto de variables}
}
\description{
Checa si las variables de una base de datos son numéricas.
}

