% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafico.R
\name{graficaCP}
\alias{graficaCP}
\title{el grafico de los primeros componentes principales}
\usage{
graficaCP(cpf)
}
\arguments{
\item{cpf}{(componentes principales finales) contiene los componentes principales finales que se trabajara}
}
\value{
Devuelve una grafica de los dos componentes principales el 1 y 2
}
\description{
el grafico de los primeros componentes principales
}
\examples{
\dontrun{
# directorio de trabajo
ruta = "C:/Users/Pablo/Desktop/yazzz/acp/lecheros.csv"
#leemos los datos
df <- read.csv("lecheros.csv")[,-1]
# imprimimos los datos
df
cor(df)
#cargamos la libreria
library(acp)
##llamamos a nuestra funcion y sustituimos para sacar los componentes principales y el porcentaje
cps <- acp(X = df)
cps
##llamamos a nuestra segunda función y sustituimos, estos serian los componentes principales con los que trabajaran
cpf0 <- cpFinales(acp = cps, no_comps =3)
cpf0
##llamamos a nuestra funcion y sustituimos para graficar los dos priemros componene
graficaCP(cpf0))
}
}
