\name{polio}
\alias{polio}
\docType{data}
\title{
Polio cases in USA from Jan 1970 till Dec 1983
}
\description{
The data set contains the monthly number of cases of poliomyelitis in the United States between 1970 and 1983.
}
\usage{data(polio)}
\format{
  The dataset consists of one variable of 168 monthly observations.
  \describe{
    \item{\code{polio}}{a numeric vector}
  }
}
\source{
Zeger, S. A Regression Model for Time Series of Counts. Biometrica, Vol. 75, No 4 (Dec., 1988), pp. 621-629  
}
\examples{
data(polio)
}
\keyword{datasets}
