% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/valida.R
\name{valida}
\alias{valida}
\title{Funcion para validar variables}
\usage{
valida(base, v)
}
\arguments{
\item{base}{base de datos}

\item{v}{conjunto de variables}
}
\description{
Esta funcion valida si las variables estan en la base de datos.
}

