% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/acp.R
\name{acp}
\alias{acp}
\title{Análisis de componentes principales}
\usage{
acp(base, v)
}
\arguments{
\item{base}{base de datos}

\item{v}{conjunto de variables}
}
\description{
Análisis de componentes principales.
}

