% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acp.R
\name{acp}
\alias{acp}
\title{aplicacion de los componentes primcipales}
\usage{
acp(X)
}
\arguments{
\item{X}{(dataframe o matriz) que contenga en las columnas el número de variables a analizar.}
}
\value{
Devuelve una lista con los componentes principales y el porcentaje de varianza explicada por componente y el acumulado.
}
\description{
Obtiene los componentes principales dependiendo variables del numero de variables que se tenga en los datos.
}
\examples{
\dontrun{
# directorio de trabajo
ruta = "C:/Users/Pablo/Desktop/yazzz/acp/lecheros.csv"
#leemos los datos
df <- read.csv("lecheros.csv")[,-1]
# imprimimos los datos
df
cor(df)
#cargamos la libreria
library(acp)
##llamamos a nuestra funcion y sustituimos para sacar los componentes principales y el porcentaje
cps <- acp(X = df)
cps
}
}
